restructure EnterPassword
