+++ WalletWasabi.Fluent/Views/Dialogs/EnterPasswordView.axaml
@@ -4,7 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:c="clr-namespace:WalletWasabi.Fluent.Controls"
              mc:Ignorable="d" d:DesignWidth="600" d:DesignHeight="500"
-             x:Class="WalletWasabi.Fluent.AddWallet.Common.EnterPasswordView"
+             x:Class="WalletWasabi.Fluent.Views.Dialogs.EnterPasswordView"
              xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors;assembly=WalletWasabi.Fluent"
              xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity">
   <UserControl.KeyBindings>
