+++ WalletWasabi.Fluent/Views/Dialogs/EnterPasswordView.axaml.cs
@@ -1,7 +1,8 @@
 using Avalonia.Markup.Xaml;
 using Avalonia.ReactiveUI;
+using WalletWasabi.Fluent.ViewModels.Dialogs;
 
-namespace WalletWasabi.Fluent.AddWallet.Common
+namespace WalletWasabi.Fluent.Views.Dialogs
 {
 	public class EnterPasswordView : ReactiveUserControl<EnterPasswordViewModel>
 	{
